feature[test]: refactoring issue #353
