--- mockk/common/src/test/kotlin/io/mockk/gh/Issue353Test.kt
@@ -1,32 +1,0 @@
-package io.mockk.gh
-
-import io.mockk.Runs
-import io.mockk.every
-import io.mockk.just
-import io.mockk.mockk
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class Issue353Test {
-
-    @Test
-    fun testNullableCapture() {
-        class Mock {
-            fun call(unused: String?) {
-                println(unused)
-            }
-        }
-
-        val list = mutableListOf<String?>()
-        val test: Mock = mockk {
-            every { call(captureNullable(list)) } just Runs
-        }
-
-        val args = listOf("One", "Two", "Three", null)
-        for (arg in args) {
-            test.call(arg)
-        }
-
-        assertEquals(args, list)
-    }
-}
